Extract ``_prepare_docstrings()``
